"8:05am. I am up. Let me do some experimenting. Got a long message by Zaid Ajaj that I need to go through.
8:10am. First of all, how do Fable do uninitialized values. Are they `undefined` or `null`?

```
let q = Array.zeroCreate 10
printfn "%i" q.[0]
```

```
(function () {
  const arg10 = q[0] | 0;
  const clo1 = (0, _String.toConsole)((0, _String.printf)("%i"));
  clo1(arg10);
})();
```

I see, so that is how it is.

```
let q = Array.zeroCreate 10
let a,b = q.[0]
printfn "%i,%i" a b
```

What about this?

```
const q = (0, _Array.fill)(new Array(10), 0, 10, null);
```

This is how it inits it. I admit, I do not understand this strange way of applying things, but I'll interpret the above as being the same thing as `_Array.fill(new Array(10), 0, 10, null);`

So it specifically nulls its fields.

9:10am. "> Actually, Fable does the exact opposite

I had to some time to think during the night of the full implications of what changing the representation of `option` would entail. What Javascript uses `undefined` is what .NET uses `null` for - uninitialized values. Changing the representation of `option` is one thing, but if Fable wants to be truly JS idiomatic rather than .NET idiomatic, its uninitialized values would need to be `undefined`.

As they are one of the rare ways to get accidental nulls in F#, right now I am looking into how uninitialized arrays are being handled. It seems special code is generated so that arrays of value types are set to zero, and for reference types, they are specially inited to `null`. I am guessing that null classes have special handling code so that their undeclared references are set to `null` rather than `undefined`.

No way is this a natural thing to do in JS.

I can think a few more ways in which Fable is .NET idiomatic - its arrays are fixed by default, while the JS ones have variable sizes. This is also something that needs to be imposed on purpose.

Furthermore the way the `async` monad works is not like Typescript's `async/await` - it does not work over the native `Promise`s. I'd like to have that happen and make my own builder, but I am not sure how to implement the monadic bind for `Promise`s.

> The recommendation s that you write the bindings by hand because code generated by ts2fable won't make anyone happy.

Those 10k LOC worth of bindings? I have better things to do than become the maintainer of such a large quantity of boilerplate.

For the VS Code API, I think the generated bindings are of pretty good quality - excepting the option issue here. Assuming this issue is taken care of promptly, and in the way I've suggested with the equivalent of ReasonML's `JS.null` and `JS.null_undefined` being added to the standard library, my intention is to open as issue for `ts2fable` so it takes advantage of them. I'd like this to be standardized so we can converge on a common idiom rather than have this be broken when it inevitably gets added to the library.

> (Those examples)

I do not want to use ugly hacks for what should be one of the most essential interop concerns that the language should have solved years ago.

Look, I like both F# and functional programming, but I am not going to bend over backwards to please the Fable compiler. If I am going to write Fable as opposed to F#, of course I'll expect it to handle optional function arguments the same way JS and TS would. Why would somebody even use Fable unless he wanted to interop with the JS world in a functional manner? I can't imagine any reason at all why I should accept the current state of affairs where Fable treats undefined variables as nulls as if it were .NET code. I don't need this extra work."

Now I am arguing with the Elmish book guy. The ideal way this could go is if the language devs all do what I say, but if I am going to encounter resistance - which I realistically will in 99% of timelines, I would be better of switching to Bucklescript just like I had started yesterday. The human nature is to look for ways to accommodate the existing poor design.

Fable devs only have themselves to blame for not putting more thought into interop. They botched the most important thing of all.

Oh crap, what did they do? If there is one place where they should have been careful it is this. This mistake is a deal breaker. It pretty much ruins the language.

9:25am. Forget this. I won't get my way in Fable. Even if they do all that I want, if this whole thing will take more than a week, I'd still be better off switching. Somewhat ironically, I will write the bindings, but it won't be in Fable.

I really can't imagine the devs dropping everything in a sudden burst of enlightenment in order to implement these vital changes.

Now let me move on...

9:30am. For the next hour, let me just finish reading the BS docs. Where was I?

https://bucklescript.github.io/docs/en/object-2

9:45am.

"Why [bs.meth]? Why not just call it directly? A JS object might carry around a reference to this, and infamously, what this points to can change. OCaml/BuckleScript functions are curried by default; this means that if you intentionally curry say, by the time you fully apply it, the this context could be wrong:"

Let me just read this and then I will go for that break.

10am. https://bucklescript.github.io/docs/en/class#new

Let me stop here for a bit.

10:20am. Let me resume.

10:30am. https://bucklescript.github.io/docs/en/class#bind-to-js-classes

The way Ocaml classes work is so strange. I can't figure out what all this is for. The syntax feels awkward.

10:45am. https://bucklescript.github.io/docs/en/function#trick-2-polymorphic-variant-bsunwrap

This is a surprise. It turns out BS does have erased unions.

https://bucklescript.github.io/docs/en/function#constrain-arguments-better

It has enumerated strings. Actually these two features were something I was considering sticking with Fable for, but it turns out that is unnecessary.

10:55am. https://bucklescript.github.io/docs/en/function#special-case-event-listeners

So far, BS is displaying really good sense for language interop. Whereas I've overestimated Fable, I find that I've been underestimating BS.

11:10am. https://bucklescript.github.io/docs/en/function#pitfall

I am going to have to learn this as I go along. No way can I grok this just by reading.

Let me take a little break again here.

11:35am. Let me resume.

https://bucklescript.github.io/docs/en/function#modeling-this-based-callbacks

I am not sure that I get this. I am going to have to figure it out as I go along.

11:40am. https://bucklescript.github.io/docs/en/return-value-wrapping

This is good sense.

"When the return type is unit: the compiler will append its return value with an OCaml unit literal to make sure it does return unit. Its main purpose is to make the user consume FFI in idiomatic OCaml code, the cost is very very small and the compiler will do smart optimizations to remove it when the returned value is not used (mostly likely)."

12pm. https://bucklescript.github.io/docs/en/exceptions

There is something that bothers me. Isn't there a way to import a module of externs wholesale? Similar to how Fable does its thing with the interfaces? I won't have to prefix every method with its path manually will I?

12:05pm.

```
let handlePromiseFailure = function [@bs.open]
  | Not_found ->
    Js.log "Not found";
    Js.Promise.resolve ()
```

What is this [@bs.open]. Does it just return `Some` if the pattern matches, but `None` if it fails?

12:20pm. https://bucklescript.github.io/docs/en/generate-converters-accessors#usage

Let me stop here so I can have breakfast. Let me read the reply by Maxime, but as a matter of policy, I will try to detatch myself emotionally from that issue. The ball is in their court now.

12:30pm. It is just on the `async` question.

Let me stop here for a while."