"4:05pm. I am up. I should be working on Spiral. Let me try to ease myself into the current sample.
4:10pm. Not yet. I still feel like just dwelling on things. So let me do just that. I am tired of bed, so let me just sit here.

6:45pm. So I ended up thinking all this time in bed.

6:50pm. I enter this kind of mind state every once in a while. I stop to think at the meta level. What are my principles and what am I trying to accomplish. This plugin adventure taking so long without even starting has definitely been enough to unsettle me. So I'll take some time to resettle.

I have various thought going through my head.

One with respect to ML - I really do not want to pray for a mathematical miracle in the understanding of intelligence.

Over the past decade there have been no real breakthroughs and yet AI advanced just fine through steady and incremental optimization. Maybe the pieces really are out there and just haven't been put together. Maybe the OpenAI guys are right, but just haven't applied themselves correctly.

With regards to Spiral and programming in general - I keep wondering whether there is a point in optimizing my skills like this. In a way, these meaningless sample refactorings are an image of the past 5 years. Is there a point to this? I think that there is. In fact doing it the opposite way is meaningless.

6:55pm. If I keep improving eventually my skills will grow into having real world use.

Spiral - it started as a crude codegen, but eventually everything became about it. Because I do not know how to hit human level AI off the bat, I took a conservative approach and focused on my own skills instead. Doing it this way will eventually prove to be right.

Every programmer wants to be good, but you cannot really improve your skills without improving your tools. I need to get back to it. I need to focus on these samples and finish them. Typescript will rot my brain.

The way I programmed in 2018 where I worked on ML and Spiral side by side was right. I should get back to it. Once I get top down typing back online, I should have a much smoother workflow. That should be fun.

If learning is program search then optimizing the way I program is surely the way to go.

7:20pm. Just because I do not know how to win is no reason to stop fighting.

7:25pm. In terms of how long humans have existed, modern programming is at most 30 years old. I think Scheme only invented lexical scope in the 80s.

But that is really stretching it. In the 80s and even the 90s people were still using assembly because high level languages (such as C) were too slow. Modern programming is in the now.

It is not like math which is seemingly stuck in the 17th century.

The art of programming is still under development and won't happen separately from AI development. If something like AI was possible to figure out separately from computation, mathematicians would have long done it.

I need to become to programming what those old guys in wigs were to mathematics. I weave all the threads together into a coherent whole. The current state of math is like Javascript - the turd on the face of its respective domain. So science and programming respectively.

7:35pm. The new of hardware will nearly hit and I am still twiddling my thumbs here. Tomorrow, I will focus and deal with the sample I should have today.

I'll only have to bear Typescript for a little while longer. Doing the client side really won't take much code or effort at all."