"8:30am. I am up. Today I am going to do it properly. Let me chill just a bit and then I will start.
8:45am. Let me start. I can slack later.

I just have to level up a bit more. 2020 will be a year to just grind these skills. Once I set things up, I will be able to do AI properly in the sense of making continuous improvements to my agents.

RL is not special. Once things are done properly, it will go through the same way supervised learning did. I need to get to that point.

Damn, I really, really wish I could have been programming this in a functional language. TS is low level trash, even though it is far better than vanilla JS.

Just a little bit more and then I will be able to move on.

Let me get `semantic-tokens-sample` to work. I tried running it yesterday, but it failed. What I am going to do is get rid of the proposed API stuff and do it directly using the latest version of VS Code which has the feature integrated.

First thought let me study the code.

8:55am. Let me check out one thing - during the night I've been wondering why the KFAC update is squared. I had that written out in that draft paper, but redoing the thing in my head after all this time shows that it makes no sense.

```
c = a * b = 4, grad = 1
a = 1, b = 1 -> a' = 2, b = 1, c' = 2
a = 0.5, b = 2 -> a' = 2.5, b = 2, c' = 5
a = 0.25, b = 4 -> a' = 4.25, b = 4, c' = 17
```

What the hell is that `c = a * b = 4`. Clearly from the examples it should be 1.

```
c = a * b, grad = 1, a' = a + grad * b
a = 1, b = 1, c = 1 -> a' = 2, b = 1, c' = 2
a = 0.5, b = 2, c = 1 -> a' = 2.5, b = 2, c' = 5
a = 0.25, b = 4, c = 1 -> a' = 4.25, b = 4, c' = 17
```

9:05am.

///

Suppose `b` is a 1d whitening matrix, `a` is the weight matrix and `c` is the two combined. What is the update for different values of `b`? Considering that the update rule for `a` is `a' = a + gradient * b` then...

```
c = a * b, grad = 1, a' = a + grad * b
a = 1, b = 1, c = 1 -> a' = 2, b = 1, c' = 2
a = 0.5, b = 2, c = 1 -> a' = 2.5, b = 2, c' = 5
a = 0.25, b = 4, c = 1 -> a' = 4.25, b = 4, c' = 17
```

Now suppose `a` and `b` are not separate as in PRONG, but we have `a * b` and `b` like in KFAC. Then updating the weight matrix does not produce correct results using the `c' = c + gradient * b` rule.

```
c = a * b
c = 1 -> b = 1 -> c' = 2
c = 1 -> b = 2 -> c' = 3 // Should be 5.
c = 1 -> b = 4 -> c' = 5 // Should be 17.
```

As shown above the right relationship implies that right quantity for multiplying the gradient is the square of `b`. `a * b` after the update would be `(a + gradient * b) * b = a * b + gradient * b * b`. Meaning `b * b` is what should be used here to add to the gradient just as the first example implies.

The proof involving multidimensional matrices rather than scalars is similar.

///

Let me modify this a bit. I get it now. I completely forgot about this.

9:10am. Yeah, I see how `b` gets double counted. But this whole exercise shows that using scalars to transmit distance is not the correct way to go. Instead it should be frequency. If you want to make large moves, you need to repeat the update instead.

The mid point between backprop and reversibility would be some kind of frequency update scheme and it would have the best points of both. It might be more computationally expensive than either.

I've been thinking about these lines for a while, but this is the first time I've actually said it out loud. The way to go around instability issues in RL due to variance of rewards is to use frequency based updates.

I don't really know how to do this though...

9:15am. It would be tough to do in RNNs.

...There are various ways though.

Literally every I can think of would greatly be benefited by having weights on chips though. As I thought, to really get anywhere, I will have to ditch the GPUs.

9:20am. The way to derive frequency based updates from regular gradient updates is to imagine dividing the learning rate `l` by some positive value `n` and counter that by updating the net by `n` number of times.

Regular gradient updates are when `n` is 1. As `n` goes to infinity, you get complely smooth updates with no noise introduced by the optimization process.

By that I mean, it could be assumed that SGD updates make discontinouus jumps which introduces noise into the optimization process. This actually helps get it around the saddle points and local minima, but if the jumps are too large then the noise swamps the process and you get instability. As `n` goes to infinity both the benefit and disadvantage of noise goes away.

9:25am. Even ignoring the noise from large updates, you still need ensembles as deep nets are nonlinear. With nonlinearity, you get switching behavior internally and even a small update can lead to large policy moves. To fight that, it would be best to use ensembles.

9:35am. Yeah, this story is coherent. There are problems with making it work at scale, but it could be done.

I still do not understand though how spiking nets could work. What sort of update scheme would be good when the nets are discontinuous completely?

That will all have to be figured out.

I should be able to do it. This kind of engineering is right up my alley. I just need to set everything up.

Forget this for now. I need to get back to studying the sample code. That is my mission at this moment. The rest can wait for later."