"10:35am. I got very little sleep tonight. I went to bed, but my brain just would not turn off. Amongst other things, I've been thinking about a new feature for heap records, but I decided against it. I am going to make unpacking them manually by using `indiv` or individual field accesses a requirement. This was my original intention.
10:40am. I am checking out extensions names in the VS Code marketplace. `.spr`, `.spa` and `.spl` are taken, but `.spi` is empty.

So `.spi` will it be. `.spi` will be the extension name for Spiral modules.

10:50am. Regrets...I've really made so many in my Spiral programming. The language was not designed right. I fell too much in love in its power to notice that I should not be putting my full strength into every swing. I was just too much into exploring a new world to really notice how low my productivity really was. Now, it has been thoroughly explored and I am a much wiser person.

It has been an excellent exercise, but I sure am sick of that by now. There has been enough exercise. This time I will aim for my goal with exquisite skill.

10:55am. The only thing I did right in the last 3 months is getting to intermediate stage with reactive extensions. Skill at that will last me and give me confidence. Well, and finally breaking through with a crappy, but popular language like Typescript.

It feels like math all over again. Having to eat shit is something I have real difficulty with.

11am. I am not really slacking right now. I am in a high tension state and yet not moving. It is not like I am trying to decide what I should first either.

11:05am. It is not like I lack desire - I will definitely start this soon.

11:10am. I am just looking inwards and taping into my purpose. From January 2019 to now, things have been all wrong. The way Spiral v0.2 is wrong as well. I was just so pissed at those math papers that I dropped them like a sack of shit.

This is not the right mindset. You go forward because you believe you can win. Because you believe that the strength and the weight of your experience will allow you to prosper.

11:15am. The reason to make a language is because you believe it will allow to create even greater vistas of power.

I need to believe - that the ensembles really are the big thing for the next wave. That I need Spiral to tap into the power of those neurochips.

Reasoning points in that direction, but I really channeled a lot of will back in 2018 only it to have it not be enough. It has left me burned. I need less skill, and more youthful vigor. I guess beyond a certain point in life, you always feel like you are old. I might end up living forever if I win, but at almost 33 it feels like I am an old man.

It is not like I want to sacrifice my present for my future.

If the future wants me to get there, then it should endeavor to make the present interesting.

11:20am. I really do want to programming in Spiral v0.2. That language will really be great.

11:25am. Everything in the way the last 5 years have unfolded deserves contempt. It would be so wonderful to continually climb towards greater capabilities in the field of RL, in the language of your choosing.

It is really a simple dream that I have. Though its implications are sinister, the day to day work should be plain.

The ML field is completely incompetent and yet I cannot beat it. Should that be funny or tragic?

10:30am. Looking inside myself, I find that my desire to program in v0.2 is quite real. This more than anything else is what will drive me forward from here.

The last 5 years have been long. I am not a beginer anymore. I've always had talent, but not the wealth of skill and experience to this degree.

There is no way the next 5 years will play out the same way the last 5 have.

Even though typechecking and doing editor support should be a significant challenge - just consider how long it took me to get Spiral into an usable state to begin with. But it honestly feels to me now that those grand challenges are just another encounter that I will pass through without much trouble.

A naive heuristic is telling me that this will be hard, but my programming intuition is telling me that it won't.

11:40am. I want to prove...that failure and lack of progress is not something that should accompany doing ML and PL side by side. That this madness can by curtailed and channeled productively.

11:45am. There is no better cure for insanity than victory. At that point it becomes genius. Like playing a game one is good at, there is a certain calmness in skill.

11:50am. There has never been a need for madness, or doubt and fear. Should I not know by now what most of the principles of ML are?

Scaling them will be harsh and for that I need the right tools.

11:55am. Let me take a break here.

Today I am too strained to really program or do anything else for that matter. Just like two days ago, I will take the time to dwell and collect my threads of thought. Let this inner war rage to completion and then I will start my charge once more. Not because I am out of options, but because I know what I need to do.

The first wave of ML is over, but the second one will start in the nearby future. When that happens I will be at my finest."