"1:25pm. https://mangadex.org/chapter/793187/11
Rebuild World is something I'll look forward to from here. It has my interest now. I like its atmosphere.

2:15pm. Done with chores. Let me read another chapter or so of this and then I'll resume the Hopac guide.

2:55pm. That was fun. I'll leave the latest two for later. Let me resume the Hopac guide.

3:05pm.

```
let rec fib n = Job.delay <| fun () ->
  if n < 2L then
    Job.result n
  else
    fib (n-2L) <&> fib (n-1L) >>- fun (x, y) ->
    x + y

let x = Diagnostics.Stopwatch.StartNew()
printfn "%i in %A" (run (fib 17L)) x.Elapsed
```

With `fib 18L` an exception gets thrown when using `<&>` so I can't really test the speedup compared to `<*>`.

The exception is so long that it does not fit in the terminal.

Let me do it in a debugger.

...Yeah, it is stack overflow. Ok.

3:10pm. I am still not sure what `Delay` exactly does. I guess I'll leave that for later.

3:15pm. Just why is VS still so bad at giving me the types of operators.

https://github.com/Hopac/Hopac/blob/master/Docs/Programming.md#programming-with-alternatives

Let me move to this section. I'll skim the merge sort stuff.

3:20pm. https://www.cambridge.org/us/academic/subjects/computer-science/distributed-networked-and-mobile-computing/concurrent-programming-ml

Since I am 2/3rds of the way in already let me get this. I'll also get something on .NET concurrency. I saw a book by Daffy namedropped in the past.

It is from 1999. That is pretty old.

There is a .NET concurrency book by Terrell.

> Of the documentation you can find it is pretty shaky, however, for more information on APM, see Jeffery Richter's brilliant book CLR via C# or Joe Duffy's comprehensive Concurrent Programming on Windows.

So this is what it was.

...Ah, forget it. Let me just get the book by Terrell.

https://docs.microsoft.com/en-us/dotnet/standard/async

Here APM is mentioned, but that is legacy by now. Still, even though I have a grasp of the `async` monad, there are still rough spots so I'll keep Terrell's book as a reference.

https://docs.microsoft.com/en-us/dotnet/standard/parallel-programming/data-structures-for-parallel-programming

I know nothing about this - well, apart to avoid it.

...Let me also get the book by Richter. It is 900 pages, so I am not going to be reading it, but I'll take a peek. Whether it be reactive combinators or channeling combinators, I want to know how to implement them under the hood. Just like I would not want to do UIs directly, so I would not use primitives directly, but for me to consider myself skilled, I should know to implement the abstractions I am using under the hood.

3:40pm. Damn, I got the `.epub`. I am not sure if the reader I am using is really shitty, but I haven't been able to find anything good for those. They are only suitable for fiction.

...Yeah, the PDF is much nicer. Well worth the extra hard drive space.

3:45pm. Judging by the TOC, I already know most of the stuff in it, but there is a bunch I might be interested in. I'll leave it for later. Let me get back to the Hopac guide. Let me get that out of the way first.

3:50pm. https://github.com/Hopac/Hopac/blob/master/Docs/Programming.md#choose-and-after

He'll sketch out a GUI example here. This stuff is interesting - it does feel more elegant than reactive combinators. Maybe I should be using this. Though I do not understand it well enough to determine whether they could be a wholesale replacement for reactive combinators.

4:10pm. https://www.demystifyfp.com/fsharp/blog/concurrent-programming-in-fsharp-using-hopac-part-1/
https://news.ycombinator.com/item?id=12571904

I am looking up some Hopac threads on HN. Interesting comments.

4:25pm. I am super distracted right now. I should not have gone for those books or looked at HN threads. I am just daydreaming now.

Focus me, focus. Finish the guide. There is like only a fifth left to go.

4:45pm. I am really skimming this last section. It feels like this part of the guide is going too fast for me. I really need a more in depth guide to this. I am not sure what to think about this.

```fsharp
let start = Job.delay <| fun () ->
  let locks = Dictionary<int64, Queue<Ch<unit> * Alt<unit>>>()
  let s = {unique = 0L; reqCh = Ch ()}
  (Job.server << Job.forever)
   (Ch.take s.reqCh >>= function
     | Acquire (lock, replyCh, abortAlt) ->
       match locks.TryGetValue lock with
        | (true, pending) ->
          pending.Enqueue (replyCh, abortAlt)
          Job.unit ()
        | _ ->
          Alt.choose [Ch.give replyCh () ^-> fun () ->
                        locks.Add (lock, Queue<_>())
                      abortAlt]
     | Release lock ->
       match locks.TryGetValue lock with
        | (true, pending) ->
          let rec assign () =
            if 0 = pending.Count then
              locks.Remove lock |> ignore
              Job.unit ()
            else
              let (replyCh, abortAlt) = pending.Dequeue ()
              Alt.choose [Ch.give replyCh ()
                          abortAlt ^=> assign]
          assign ()
        | _ ->
          // We just ignore the erroneous release request
          Job.unit ()) >>-. s
```

I really don't want to think about this right now. Let me finish the guide. I am satisfied with understanding the first half. I'll go for the Concurrent ML book next. It feels like this part is just covering it; like a proof that whatever the language can do, the F# library can do it as well. Because in isolation, the above example is just so poorly motivated.

4:55pm. The Hopac library is so extensive. It is a real hidden gem. Let me check out the book.

The guide was roughly 40-50 pages long, but the book is 325. So I'll take my time studying it.

5:10pm. Actually, it is interesting that Rx came 10 years after the work done on CML. I'd figured that Rx would have been simple to produce, but I guess not.

5:25pm. 19/325.

> Interactive systems, such as graphical user interfaces and window managers, are the primary motivation of much of the work described in this book, and the author believes that they are one of the most important application areas for concurrent programming.

Interesting that he specifically singled out UIs. I was thinking that that Rx was a better fit, but who knows.

Well, I do not particularly mind sinking more time into this. My sense is that I'll continue using Rx, but I might as well get familiar with other options.

5:40pm. Let me take a break here. I am not done yet.

5:50pm. Let me resume. It is time for the last spurt. I'll do it for another 30m and then I'll call it a day. Since I am just reading books today, I am definitely not in the mood for an 8pm session.

5:55pm. 31/325. Maybe I should just stop now. I can't really properly focus on this. It is kind of boring, but I should really pay attention to it.

6pm. 32/325. Ah, so this is the producer consumer process. There were some wildly out of place functions in the F# Rx bindings that I had no idea what they were supposed to be doing.

...Well, this is just a brief description.

Let me stop here. I do not feel like going any further. I know I said I would do 30m, but forget that.

This is not like programming where I get into it. When I am in passive mode my limit is 6pm.

What I will do tomorrow is continue this book. Even if I have to skim it, I'll read it cover to cover to get a gist of it. Right now, I am still not sold on the utility of this versus Rx and I have so much on my plate, so I do not want to spend more than a few days on this.

I just want to learn enough to understand what it would be useful for."