"9:35pm. I took a little break. And I think I figured out everything.
It is not that the thread wakes up. It is the other side that wakes it up.

You can get the efficiency that you want if the other side rather than just filling up the buffer also knew who the worker waiting on it is. Then it can knock him awake. And then the worker if needed can knock the main thread awake in case it is dormant. I see it now.

I could not understand this before, but now I clearly have a model where a queue gets passed around through the CPS'd program.

Even with a single thread you could get this asynchrony simply by having the channel execute the other's side's continuation.

9:45pm. Ok, this is great. Obviously, I do not get the exact details of how to do this, but I've reached it.

I have the general model of CML/Hopac in mind.

With this if I wanted to, I could come up with a rough version of Hopac on my own now.

9:50pm. I've been thinking of the PC buffer all wrong now.

The way the book presented it, it is a full/empty slot. Something like a concurrent `'a option`.

But the right way to think about it would be...

```
{full : ('a -> unit) list; empty : ('a -> unit) list}
```

Or better yet...

```
{full_or_empty : ('a -> unit) list; is_full : bool}
```

Basically, it only ever possible to wait to add or to remove from the buffer. With this framing you do not even need the scheduler, you can just have the side that is active trigger the other and remove it from the list.

Actually...

```
{full_or_empty : ('a -> unit) Queue; is_full : bool}
```

The `Queue` would be the most suitable for this.

10pm. I could not see it at first, but this formulation simplifies a vast amount and it makes it look a lot closer to reactive combinators.

10:05pm. Actually, I am a bit wrong with the types here. It is 10pm so I am not thinking completely straight.

```
{on_receive : ('a -> unit) Queue; on_send : (unit -> 'a) Queue}
```

This is more likely the correct type. So, if `on_receive` is not empty the sender would trigger it, otherwise it would enqeue itself on `on_send`. And vice versa for the other side.

This is really brilliant and simplifies pretty much everything.

One thing I do not understand is how to dispatch on a particular thread. I'd need that for this to work, but it should definitely be doable.

10:15pm. Right now I am thinking about message boxes, `IVar`s and `MVar`s. The message box is just a straightforward function `'a -> ()`. `IVar` is...

```
{on_receive : ('a -> unit) Queue; on_send : (unit -> 'a) option}
```

It is actually less complicated that the seemingly more primitive channel.

`MVar` would be...

```fsharp
type 'a QueueOrValue = Queue of ('a -> unit) Queue | Value of 'a
{on_receive : 'a QueueOrValue}
```

10:20am. Something like that. My brain is not working properly at this point in the day and I am getting confused at what is what. But with MVars, you only need a queue for the receivers when the `MVar` is uninitialized.

10:25pm. It terms of implementation complexity: channel > ivar > mvar > message box. It is interesting, that this is nearly the opposite of their power. The channel is the one that is the most restricted of the bunch.

10:30pm. Indeed, with this scheme you definitely would not need to manage subscriptions like with reactive combinators.

10:35pm. Let me go once more from the top. This time I'll use union types.

Channel:

```fsharp
{receivers_waiting_on_senders : ('a -> unit) Queue; senders_waiting_on_receivers : (unit -> 'a) Queue}
```

IVar:

```fsharp
{receivers_waiting_on_senders : ('a -> unit) Queue} | {sender_waiting_on_receiver : (unit -> 'a)}
```

MVar:

```fsharp
{receivers_waiting_on_sender : ('a -> unit) Queue} | {value : 'a}
```

MessageBox:

```fsharp
{receiver_waiting_on_sender : 'a -> unit}
```

11:50pm. This should be fine. Obviously this is pseudocode. Though I did it in one of the examples, I do not want to use union types for everything.

Let me make a very late commit and then I'll get dinner. This should be enough to afix the idea. You could read the whole source code, all 5k lines of it, and you would not realize the above forms from it."