"8:10pm. I want to study the Hopac API just to alleviate my doubts, but this is really not the time for it. Let me slack for a bit and then I will go to bed.
8:45pm. https://github.com/Hopac/Hopac/blob/master/Examples/ToDo/ToDo.fs

```
// The main problem with this example is that this doesn't really contain
// anything where choice streams work significantly better than Rx, such as
// asynchronous http requests, polling, streams requiring backpressure, or
// initialization of streams that would require hacks like `Publish-Connect` in
// Rx.  Although it doesn't really make any difference in this example, the
// `keepPreceding1` combinator is an example of something that is not supported
// by .Net Rx.  See comment below.
```

5/24/2020

7:55am. I am up. Yesterday I praised the reactive combinators to death, but it is not a given that I will pick them over Hopac. As I said, UIs are one thing, but I want something that will give me the most power and flexibility when it comes to doing the language support.

I have in my mind chains of modules. And I want to do this efficiently - I want to make it so that when the user moves from file to file, that the attention being put is being used effectively.

8:10am. Though I could move to the next example, today I want to do more experimentation. I do want to look into how pull streams work. I also want to investigate `AutoConnect`. Does that thing only manage connections? If so, how would I dispose of the whole chain?

I won't hurry this. I know that I have a lot to learn before I am ready to start.

Now, let me slack for a while and then I'll get on with it.

8:20am. Agh, I cannot wait.

```
    let a = Subject.broadcast
    let a' = a.Select((*) 2).Publish().AutoConnect()
```

...No, this won't work. But I am wondering what that second optional argument is.

http://reactivex.io/RxJava/2.x/javadoc/io/reactivex/Observable.html#cache--

This is RxJava.

8:25am. http://reactivex.io/RxJava/2.x/javadoc/io/reactivex/observables/ConnectableObservable.html#autoConnect--

For what I want, I do need a persistent connection. Still, the observable chain might have resources I want to dispose when it is entirely done.

8:30am. I need to play a bit. Let me see if autoconnects manages connections at least. Regarding the disposable, I have an idea.

```
    let s = Subject.broadcast
    let a = s.Publish().AutoConnect().Do(printfn "a got %A")
    let b = a.Publish().AutoConnect().Do(printfn "b got %A")
    let c = b.Publish().AutoConnect().Do(printfn "c got %A")
    let _ =
        use __ = c.Subscribe()
        s.OnNext(22)
    let _ =
        use __ = a.Subscribe()
        s.OnNext(11)
    0
```
```
a got 22
b got 22
c got 22
a got 11
b got 11
a got 11
```

This is completely useless.

I get that you won't manage the inner subscription, but at least get rid of the connection when there are no subscribers.

8:55am. At this point though, I am getting too deeply involved with managing subscriptions.

Had to take a short break. I am wracking my brain over this, and I can't escape the sense that I am on the wrong path.

...Decided to go look at the extension options and I see that there is a recheck entire document on save option. That needs to be off.

...

Now, Hopac. Rx is good for UIs, but it is not going to get me all the way to the end after all. Past a certain point, the subscription burden becomes a weakness.

I am going to have to look into Hopac after all.

As time consuming as this will be, I am going to have to try doing UIs pull based fashion just to see where that gets me.

9am. I just can't do this quickly. I can't master concurrency just like that.

https://vorpus.org/blog/notes-on-structured-concurrency-or-go-statement-considered-harmful/

Hmmm, this is interesting.

9:30am. https://vorpus.org/blog/some-thoughts-on-asynchronous-api-design-in-a-post-asyncawait-world/

I'll read this later. Forget about this for now.

9:55am. Ok, focus me. Leave that Shuten thread from two days ago alone.

It is time for Hopac. I need practice, practice, practice. I need hard training to break through this current hurdle.

Right now, I feel like my reactive (push) combinator skills are pretty good. But I need to master the other side as well. Only at that point will I be complete.

10am. Maybe there is way to do it using Rx, but it is not as straightforward as I would prefer it to be.

```
        Disposable.Create(fun () ->
            let pollers, tasks = Array.unzip l
            pollers |> Array.iter (fun (x : NetMQPoller) -> x.Stop())
            Task.WaitAll(tasks)
            tasks |> Array.iter (fun x -> x.Dispose())
            pollers |> Array.iter (fun (x : NetMQPoller) -> x.Dispose())
            )
```

Let me dispose of the pollers as well.

10:05am. Actually, I can in fact see the way to do it using Rx. Instead of the usual chaining, I would have to pass the observable in as the scan's argument.

Which is convoluted. It is no wonder I want to try the pull version.

Let me do it. Let's see where it gets me.

I need to make up my mind and just do it. It is finally time to get some Hopac experience.

This will no doubt take me a few days. Maybe an entire week, but it will be worth it.

How about I aim for the button stacking example first and then make my way from there?

10:35am. Uah...

I have no idea, where to even start with this. I can see how doing that file chain would be easier using the pull model, but here I am trying to do an UI library.

10:45am. I am really drawing a blank here.

10:55am. Shit.

Instead of getting ideas about how to do this with Hopac, I am pitting the two in my head and Rx is coming on top. I am figuring out how to do the observable chain in a natural and efficient manner. It is all coming together to me.

11:05am. No wait, no shit. I want servers to have persistent state even when they are unsubscribed to.

The key factor here seems to be whether I should use disposable on GC'd state. If I have...

```
                    num_buttons.SelectMany(fun x ->
                        stack_panel [
                            do' <| fun pan -> pan.Background <- Brushes.Red; pan.Margin <- Thickness 20.0
                            children <| create_buttons x
                            ]
                        )
```

Oh, this might be bad.

I do not think this will really unsubscribe from the old one.

```
                    num_buttons |> Observable.switchMap(fun x ->
                        stack_panel [
                            do' <| fun pan -> pan.Background <- Brushes.Red; pan.Margin <- Thickness 20.0
                            children <| create_buttons x
                            ]
                        )
```

Let me try this.

11:10am. Yeah, it works.

11:20am. Let me take a short break.

11:45am. I'll segue that into a big break."